Expands functionality to reference tabs by hrefs as well as numerical indexes.
Solves jQuery UI Bug #3171.

Important where you need to work with runtime-generated tabs which have a random order, and may be deleted at any time. Working with numerical indexes is problematic in this situation especially when a tab in the middle must be deleted, reindexing all subsequent tabs.

It is also a convenience option to the developers which would prefer to work with the hrefs.

It is safe because each href points to an ID, which per spec must be unique.

A single internal function _getIndex() is used to parse the index variable and convert it to a numerical index if needed. If the href is not found it returns NaN. It also does a little bit of sanitation in numbers it receives. Always returns a typeof number.

a one-liner index = this._getIndex(index); is all that is needed on other functions to benefit from this.

Functions altered: enable, disable, select, remove and load.